[IMP] wingpanel-slim: default spacing is now 60
Considering the new Gnome headerbars that are
used in Freya, this is not enough to avoid
all overlapping. However, it is enough to
avoid overlapping with the settings menu in the
headerbar, which is the best I could find while
still keeping the original placement feeling.

There is actually no value which will avoid all
overlapping, thus auto-hide will probably be
mandatory to use it day-to-day, unless you
don't really use any app with headerbars, or
you don't mind the overlapping. I intend to add
a Hide on Maximize feature which would help
with this issue.